Made a Simple Renderer, Updates to Application
- Added SDL
- Added Imgui
- Made a simple Renderer class to draw a few primitives
- Application now is a base class, so that I can easily modify the Update() to test things.
- The 'Platform' class is gone. It was confusing as to who owned what between the Platform and the Application, so I simplified and moved everything into the Application.
- NES_WINDOW_API and NES_RENDER_API macros were added to enable and disable certain implementation files.
- RenderAPI is a setting in the Project file, which will include certain projects based on that API. This was done because I wanted to switch from GLFW to SDL to get something going quickly. I plan on using Vulkan with GLFW, but that will be done in the upcoming semester.

+++ Source/Tools/ProjectGenerator/Modules/Cleaner.lua
@@ -78,6 +78,11 @@ function m.CleanSolution()
         return false;
     end
 
+    -- .idea folder (Rider)
+    if utility.DeleteFolderIfExists(core.SolutionDir .. ".idea/", ".idea") == false then
+        return false;
+    end
+
     -- .sln File.
     if utility.DeleteFileIfExists(core.SolutionDir .. core.ProjectSettings["ProjectName"] .. ".sln") == false then
         return false;
